pyproject.toml
[project]
name = "grafana-playground"
version = "0.1.0"
description = "Grafana playground with Git Sync configuration"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.2",
]
